requirements.txt
numpy
opencv-python-headless
Pillow==8.4.0
keras==2.7.0
Keras-Preprocessing==1.1.2
tensorboard==2.7.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
streamlit==1.3.0
streamlit_webrtc==0.32.0